Fix ignore platform
The ignorePlatform was being kept until the player stepped on a
different platform. This would work fine, except on the case when the
player fell from a platform X (by pressing the down key), then took a
rocket without stepping on any platform, and then came back to the
platform X. In this case the platform X would still be the ignored, and
it would be impossible to stand on it, unless stepping on another
platform first.